some of edits, doing the range feature [but still doesn't work as expected]
