+++ src/components/Summary.tsx
@@ -3,7 +3,7 @@ import classes from './Style.module.css';
 import NumbersList from "./NumbersList";
 import {Button, Typography} from "@material-ui/core";
 import {DataContext} from "../context/data-context";
-import {Link, Redirect} from 'react-router-dom';
+import {Link} from 'react-router-dom';
 
 const Summary: React.FC = () => {
     const dataContext = useContext(DataContext);
@@ -19,7 +19,6 @@ const Summary: React.FC = () => {
         </div>
         <NumbersList/>
         <Link to={{pathname:'/'}} onClick={dataContext.backToHome}><Button variant={"contained"} color={"secondary"}>Back to home page</Button></Link>
-        {!dataContext.redirects && <Redirect to={'/'}/>}
     </div>
 };
